#27 chore: 사용하지 않는 라우트 제거 및 sessionUser 타입 분리
